Merge anchor alignment functions
The two functions are basically identical in how they work.
The transposition differences at the end only change semantics;
both could be done in the same exact way.

The only difference was column/row transposition,
so I made it to resolve based on a boolean argument.

Simplifies code and allows for plot stacking (next commit).